Added conversion to kvtml, fileformat for Parley
I made some small modifications (such as utf-8, to support languages such as Turkish and Arabic) and added a conversion to the kvtml format.
After conversion (result in output/cards.kvtml) it can be useful to validate the file with xmllint. Then open the file in Parley version 4.14+ and save the file with Parley.